+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The lib folder is moved to a package - @resubaka (#30)
 - Added hooks implementation from vue-storefront as a package - @resubaka (#30)
 - Added the integration tests - @resubaka (#35)
+- Add url module - @gibkigonzo (#3942)
 
 ### Changed / Improved